+++ e2e/tests/00_setup.yaml
@@ -28,8 +28,8 @@ testcases:
   steps:
   - script: kubectl create ns goldilocks
   - script: |
-      kubectl -n goldilocks apply -f ../../hack/manifests/web/
+      kubectl -n goldilocks apply -f ../../hack/manifests/dashboard/
       kubectl -n goldilocks apply -f ../../hack/manifests/controller/
   - script: |
-      kubectl -n goldilocks wait deployment --timeout={{.timeout}} --for condition=available -l app.kubernetes.io/name=goldilocks,app.kubernetes.io/component=web
+      kubectl -n goldilocks wait deployment --timeout={{.timeout}} --for condition=available -l app.kubernetes.io/name=goldilocks,app.kubernetes.io/component=dashboard
       kubectl -n goldilocks wait deployment --timeout={{.timeout}} --for condition=available -l app.kubernetes.io/name=goldilocks,app.kubernetes.io/component=controller
