+++ hack/manifests/dashboard/clusterrole.yaml
@@ -2,7 +2,7 @@
 apiVersion: rbac.authorization.k8s.io/v1beta1
 kind: ClusterRole
 metadata:
-  name: goldilocks-web
+  name: goldilocks-dashboard
   labels:
     app: goldilocks
 rules:
