+++ hack/manifests/dashboard/clusterrolebinding.yaml
@@ -2,14 +2,14 @@
 apiVersion: rbac.authorization.k8s.io/v1beta1
 kind: ClusterRoleBinding
 metadata:
-  name: goldilocks-web
+  name: goldilocks-dashboard
   labels:
     app: goldilocks
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: goldilocks-web
+  name: goldilocks-dashboard
 subjects:
   - kind: ServiceAccount
-    name: goldilocks-web
+    name: goldilocks-dashboard
     namespace: goldilocks
