+++ hack/manifests/dashboard/deployment.yaml
@@ -2,30 +2,30 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: goldilocks-web
+  name: goldilocks-dashboard
   labels:
     app.kubernetes.io/name: goldilocks
-    app.kubernetes.io/component: web
+    app.kubernetes.io/component: dashboard
 spec:
   replicas: 2
   selector:
     matchLabels:
       app.kubernetes.io/name: goldilocks
-      app.kubernetes.io/component: web
+      app.kubernetes.io/component: dashboard
   template:
     metadata:
       labels:
         app.kubernetes.io/name: goldilocks
-        app.kubernetes.io/component: web
+        app.kubernetes.io/component: dashboard
     spec:
-      serviceAccountName: goldilocks-web
+      serviceAccountName: goldilocks-dashboard
       containers:
         - name: goldilocks
           image: "quay.io/fairwinds/goldilocks:master"
           imagePullPolicy: Always
           command:
             - /goldilocks
-            - web
+            - dashboard
             - --exclude-containers=linkerd-proxy,istio-proxy
             - -v3
           securityContext:
