+++ hack/manifests/dashboard/serviceaccount.yaml
@@ -1,6 +1,6 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: goldilocks-web
+  name: goldilocks-dashboard
   labels:
     app: goldilocks
