+++ pkg/dashboard/assets/css/check-details.css
