+++ pkg/dashboard/assets/css/fontawesome-5.7.2.css
