+++ pkg/dashboard/assets/css/prism.css
