+++ pkg/dashboard/assets/images/favicon-16x16.png
