+++ pkg/dashboard/assets/images/favicon-32x32.png
