+++ pkg/dashboard/assets/images/favicon.ico
