+++ pkg/dashboard/assets/images/fw-logo.png
