+++ pkg/dashboard/assets/js/main.js
