+++ pkg/dashboard/assets/js/prism.js
