+++ pkg/dashboard/assets/webfonts/fa-regular-400.woff2
