+++ pkg/dashboard/dashboard.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package web
+package dashboard
 
 import (
 	"net/http"
