+++ pkg/dashboard/templates/dashboard.gohtml
