+++ pkg/dashboard/templates/footer.gohtml
