+++ pkg/dashboard/templates/head.gohtml
