+++ pkg/dashboard/templates/namespace.gohtml
