+++ pkg/dashboard/templates/namespace_list.gohtml
