+++ pkg/dashboard/templates/navbar.gohtml
